// network/types.ts
export interface PackageBookingListResponse {
  id?: string;
  user?: {
    email?: string;
    profile?: {
      first_name?: string;
      last_name?: string;
      phone_number?: string;
      address?: string;
    };
  };
  expiry_date?: string;
  created_at?: string;
  selected_batch: UmrahBatch;
  selected_price?: {
    id?: string;
    package?: string;
    price?: string;
    category?: string;
    weekly_installment_fee?: string;
    monthly_installment_fee?: string;
  };
  status?: string;
  balance?: string;
  payment_plan?: string;
  payment_summary?: {
    amount_to_be_paid?: number;
    amount_paid?: number;
    balance?: number;
  };
  is_active?: boolean;
  is_card_delivered?: boolean;
  qr_code_url?: string;
  is_on_whatsapp_group?: boolean;
  card_delivered?: boolean;
}